backend/requirements.txt
fastapi
uvicorn
openai~=1.0
python-dotenv
weaviate-client~=3.26.2
tree-sitter
sentence-transformers
torch
